requirements.txt



python-dotenv

requests
selenium
beautifulsoup4

pandas
